requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
